Implement database creation and deletion, pass 1
Due to the stupidites of SQLAlchemy, the quoting of the database is
done with single-quotes instead of backticks, as MySQL demands. So we
give up and import MySQLdb, the underlying library SQLAlchemy is
using, and escape the string directly.

Other than that, this commit implements the ability to create and drop
databases, but does not yet have the checks to prevent users from
going over their quotas.